# solution/idea/openapi.yml
openapi: 3.0.3
info:
  title: Anti-fraud Service API
  version: 1.3.0
  description: |
    # Сервис антифрода

    REST API для обнаружения мошеннических транзакций с использованием настраиваемых DSL-правил.

    ## Роли пользователей
    - **ADMIN** — полный доступ: управление пользователями, правилами, просмотр всех транзакций и статистики
    - **USER** — ограниченный доступ: только свой профиль, свои транзакции, создание транзакций для себя

    ## Аутентификация (JWT)

    Используется Bearer-токен в заголовке `Authorization: Bearer <token>`.

    **Параметры JWT:**
    - Алгоритм: HS256
    - Время жизни: 1 час (3600 секунд)
    - Секрет: из переменной окружения `RANDOM_SECRET` (128 символов)

    **Payload токена:**
    ```json
    {
      "sub": "<userId>",
      "role": "USER|ADMIN",
      "iat": 1234567890,
      "exp": 1234571490
    }
    ```

    ## DSL для правил фрода

    Правила описываются на простом языке выражений.

    **Грамматика (EBNF):**
    ```
    expression = term { "OR" term }
    term       = factor { "AND" factor }
    factor     = "NOT" factor | comparison | "(" expression ")"
    comparison = field operator value
    field      = "amount" | "currency" | "merchantId" | "ipAddress" | "deviceId"
               | "user.age" | "user.region"
    operator   = ">" | ">=" | "<" | "<=" | "=" | "!="
    value      = number | string
    string     = "'" { character } "'"
...
    ```

...

    **Примеры:**
...

...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      description: |
        Сценарии:
        - 201: пользователь создан, роль USER по умолчанию, возвращается accessToken и профиль.
        - 409: email уже используется.
        - 422: не проходит валидация (email, пароль, возраст и т.д.).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
            examples:
              ok:
                summary: Успешная регистрация
                value:
                  email: user@example.com
                  password: "Qwerty_123"
                  fullName: "Ivan Ivanov"
                  region: "RU-MOW"
                  gender: MALE
                  age: 25
                  maritalStatus: SINGLE
      responses:
        '201':
          description: |
            Пользователь успешно зарегистрирован.
            Примечание: accessToken используется в Authorization: Bearer <token>.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
              examples:
                created:
                  summary: Токен + профиль
                  value:
                    accessToken: "eyJhbGciOi..."
                    expiresIn: 3600
                    user:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      email: "user@example.com"
                      fullName: "Ivan Ivanov"
                      region: "RU-MOW"
                      gender: "MALE"
                      age: 25
                      maritalStatus: "SINGLE"
                      role: "USER"
                      isActive: true
                      createdAt: "2025-12-15T10:00:00Z"
                      updatedAt: "2025-12-15T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email уже занят
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              examples:
                conflict:
                  value:
                    code: EMAIL_ALREADY_EXISTS
                    message: "Пользователь с таким email уже существует"
                    traceId: "550e8400-e29b-41d4-a716-446655440001"
                    timestamp: "2025-12-15T10:00:00Z"
                    path: "/api/v1/auth/register"
                    details:
                      field: "email"
                      value: "user@example.com"
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /auth/login:
    post:
      tags: [Auth]
      summary: Авторизация пользователя
      description: |
        Сценарии:
        - 200: корректные credentials => accessToken.
        - 401: неверный email/пароль (не раскрываем, что именно неверно).
        - 423: пользователь деактивирован (isActive=false).
        - 422: формат email/пароля невалиден.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            examples:
              ok:
                value:
                  email: "user@example.com"
                  password: "Qwerty_123"
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          $ref: '#/components/responses/Locked'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /users:
    get:
      tags: [Users]
      summary: Список пользователей
      description: |
        Только ADMIN.
        Сценарии:
        - 200: страница пользователей
        - 422: некорректные page/size
      x-roles: [ADMIN]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedUsers' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    post:
      tags: [Users]
      summary: Создание пользователя админом
      description: |
        Только ADMIN. В ответе нет токенов (создание "из админки").
        Сценарии:
        - 201: создан
        - 409: email занят
        - 422: не проходит валидация полей (пароль/роль/возраст и т.д.)
      x-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreateRequest' }
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /users/me:
    get:
      tags: [Users]
      summary: Получить профиль текущего пользователя
      description: |
        Возвращает профиль пользователя, определённого по JWT токену.
        Удобный способ получить свой профиль без знания своего ID.

        Доступно для любой роли (ADMIN, USER).
      x-roles: [ADMIN, USER]
      responses:
        '200':
          description: Профиль текущего пользователя
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Users]
      summary: Обновить профиль текущего пользователя
      description: |
        Полное обновление профиля текущего пользователя.
        Необходимо передать все поля. Чтобы очистить поле, передайте null.

        **Обязательные поля:** fullName, age, region, gender, maritalStatus
        **Поля, которые могут быть null:** age, region, gender, maritalStatus
        **Поля, которые НЕ могут быть null:** fullName

        **Ограничения для USER:**
        - Нельзя менять role — вернёт 403
        - Нельзя менять isActive — вернёт 403

        ADMIN может менять любые поля своего профиля, включая role и isActive.
      x-roles: [ADMIN, USER]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
            examples:
              fullUpdate:
                summary: Полное обновление с null
                value:
                  fullName: "Иван Иванов"
                  age: 30
                  region: "RU-SPB"
                  gender: null
                  maritalStatus: "MARRIED"
      responses:
        '200':
          description: Обновлённый профиль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Users]
      summary: Получение профиля пользователя
      description: |
        Доступ:
        - ADMIN: любой пользователь
        - USER: только собственный профиль (id == sub/id в токене)
        Сценарии:
        - 200: найден
        - 403: попытка USER прочитать чужой профиль
        - 404: не найден
      x-roles: [ADMIN, USER]
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Полное обновление профиля пользователя
      description: |
        Полное обновление профиля. Необходимо передать все поля.
        Чтобы очистить поле, передайте null.

        **Обязательные поля:** fullName, age, region, gender, maritalStatus
        **Поля, которые могут быть null:** age, region, gender, maritalStatus
        **Поля, которые НЕ могут быть null:** fullName

        **Права доступа:**
        - ADMIN: может обновлять любой профиль, включая role и isActive
        - USER: только свой профиль (id в URL == id в токене)

        **Ограничения для USER:**
        - Попытка изменить role => 403 FORBIDDEN
        - Попытка изменить isActive => 403 FORBIDDEN
        - Попытка обновить чужой профиль => 403 FORBIDDEN

        **Валидация:**
        - fullName: 2-200 символов (обязательный, не может быть null)
        - age: 18-120 или null
        - region: до 32 символов или null
        - email изменить нельзя (игнорируется, если передан)
      x-roles: [ADMIN, USER]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
            examples:
              fullUpdate:
                summary: Полное обновление с null
                value:
                  fullName: "Новое Имя"
                  age: 25
                  region: null
                  gender: "MALE"
                  maritalStatus: null
      responses:
        '200':
          description: Обновлённый профиль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    delete:
      tags: [Users]
      summary: Деактивация пользователя
      description: |
        **Важно:** Это soft-delete. Пользователь НЕ удаляется физически из базы данных,
        а помечается как неактивный (isActive = false).

        Только ADMIN может деактивировать пользователей.

        **Идемпотентность:**
        - Повторный вызов для уже деактивированного пользователя => 204 (не ошибка)
        - Несуществующий пользователь => 404

        **Последствия деактивации:**
        - POST /auth/login => 423 USER_INACTIVE
        - POST /transactions с userId деактивированного => 403 FORBIDDEN
        - Все существующие токены пользователя продолжают работать до истечения (1 час)

        **Восстановление:**
        - ADMIN может снова активировать через PUT /users/{id} с isActive=true
      x-roles: [ADMIN]
      responses:
        '204':
          description: Пользователь деактивирован
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /transactions/{id}:
    parameters:
      - $ref: '#/components/parameters/TransactionId'
    get:
      tags: [Transactions]
      summary: Получить транзакцию
      description: |
        - ADMIN: любая транзакция
        - USER: только свои
      x-roles: [ADMIN, USER]
      responses:
        '200':
          description: Транзакция + результаты правил
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionDecision' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /transactions/batch:
    post:
      tags: [Transactions]
      summary: Создать батч транзакций
      description: |
        Батч поддерживает частичный успех.

        Сценарии:
        - 201: все элементы обработаны успешно (каждый элемент может быть APPROVED или DECLINED)
        - 207: частичный успех (есть элементы, которые не созданы из-за 422/403/409 и т.п.)
        Для сопоставления используется index (0-based).
      x-roles: [ADMIN, USER]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionBatchCreateRequest' }
      responses:
        '201':
          description: Все элементы батча обработаны (без транспортных ошибок по элементам)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionBatchResult' }
        '207':
          description: Частичный успех (поэлементные ошибки/валидации)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionBatchResult' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /fraud-rules:
    get:
      tags: [FraudRules]
      summary: Список правил фрода
      description: |
        Только ADMIN.
        Сценарии:
        - 200: список правил
      x-roles: [ADMIN]
      responses:
        '200':
          description: Список правил
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FraudRule' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [FraudRules]
      summary: Создать правило фрода
      description: |
        Только ADMIN.

        Валидация:
        - name: 3..120 символов
        - dslExpression: обязателен, должен парситься по спецификации DSL
        - priority: >=1 (меньше — выше приоритет)
        Сценарии:
        - 201: создано
        - 422: невалидные поля или DSL синтаксически неверен
        - 409: правило с таким именем уже есть
      x-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FraudRuleCreateRequest' }
            examples:
              ok:
                value:
                  name: "High amount for young users"
                  description: "Блокировать транзакции > 10000 от пользователей младше 21"
                  dslExpression: "amount > 10000 AND user.age < 21"
                  enabled: true
                  priority: 5
      responses:
        '201':
          description: Правило создано
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FraudRule' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 0, default: 0 }
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          writeOnly: true
...